+++ lib/central_config/config.rb
@@ -6,11 +6,12 @@ module CentralConfig
   class Config
     attr_accessor :adapter, :error_handler, :flagr_host
 
-    def initialize(data = {}, adapter: CentralConfig::Adapters::Flagr.new)
-      @adapter = adapter
+    def initialize(data = {}, adapter: nil)
       @data = data
       @error_handler = method(:error_warning)
       @flagr_host = ENV['CENTRAL_CONFIG_FLAGR_HOST']
+
+      @adapter = adapter || default_adapter
     end
 
     def flag?(flag_name, default: false)
@@ -21,9 +22,8 @@ module CentralConfig
       block_given? ? yield(flag_name) : default
     end
 
-    def load(entity_id:, context:, flags:)
-      self.data =
-        adapter.call(config: self, entity_id: entity_id, context: context, flags: flags)
+    def load(entity_id:, context:)
+      self.data = adapter.call(entity_id: entity_id, context: context)
     rescue StandardError => exception
       error_handler.call(exception)
       self.data = {}
@@ -38,6 +38,10 @@ module CentralConfig
     end
 
     private
+
+    def default_adapter
+      CentralConfig::Adapters::Flagr.new(config: self)
+    end
 
     def error_warning(exception)
       warn "Returning empty central config due to #{exception.inspect}"